+++ inc/functions_global.php
@@ -65,6 +65,8 @@ function in_tools($page, $do)
 		return true;
 	if ($page == 'search_log')
 		return true;
+	if ($page == 'bodyweight')
+		return true;
 	if ($page == 'exercise' && $do == 'compare')
 		return true;
 	return false;